dash-bakehouse/requirements.txt
databricks-sql-connector
databricks-sdk
dash
pandas
plotly
posit-sdk
